+++ src/Prose/Parsing/BlockType.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2014, Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
+
+namespace Prose.Parsing
+{
+    public enum BlockType
+    {
+        Code = 0,
+
+        Markdown = 1,
+    }
+}