+++ src/Prose/Weavers/IWeaverEngine`1.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) 2014, Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
+
+namespace Prose.Weavers
+{
+    using System.IO;
+    using Prose.Templating;
+
+    public interface IWeaverEngine<TModel> where TModel : ITemplateModel
+    {
+        string Weave(TextReader reader, TModel model);
+    }
+}